[Minor] Types for Card component
